fix: The type of the right operand ('BigInt') isn't a subtype or a supertype of the left operand ('int?').

+++ lib/_pkg/wallet/bdk/transaction.dart
@@ -269,7 +269,7 @@ class BDKTransactions {
               if (serdBdkTx.output == null) throw 'No output object';
               final scriptPubkeyString = serdBdkTx.output
                   ?.firstWhere(
-                    (output) => output.value == amountChange,
+                    (output) => output.value == amountChange.toInt(),
                   )
                   .scriptPubkey;
 
@@ -323,7 +323,9 @@ class BDKTransactions {
             try {
               if (serdBdkTx.output == null) throw 'No output object';
               final scriptPubkeyString = serdBdkTx.output
-                  ?.firstWhere((output) => output.value == amountReceived)
+                  ?.firstWhere(
+                    (output) => output.value == amountReceived.toInt(),
+                  )
                   .scriptPubkey;
 
               if (scriptPubkeyString == null) {
